backend/requirements.txt
celery
redis
transformers
torch
torchvision
Pillow
requests
langchain
langchain-community
langchain-core
langchain-ollama
plyer
accelerate
flash-attn